Detect circular waitFor and incorrect waitFor
This throws an error now when a circular dependency is detected (it would result in a `Maximum call stack size exceeded` if not thrown or stopped in some way).

Adds a console.warn when a store action waits for a separate store action in a store that does not exist or in which the action is not handled.